pyproject.toml
[tool.poetry]
name = "realtime-voting"
version = "0.1.0"
description = "Real-time election voting system using Kafka, Spark Streaming, PostgreSQL, and Streamlit"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "config"}, {include = "models"}]

[tool.poetry.dependencies]
# Python version
python = "^3.11"

# Phase 1: Configuration and validation
pydantic = "^2.6"
python-dotenv = "^1.0"
email-validator = "^2.1"

# Data processing
pyspark = "^3.5"
pandas = "^2.1"
numpy = "^1.26"

# Kafka messaging
confluent-kafka = "^2.3"
kafka-python = "^2.0"
simplejson = "^3.19"

# Database
psycopg2-binary = "^2.9"

# Visualization and UI
streamlit = "^1.29"
streamlit-autorefresh = "^1.0"
matplotlib = "^3.8"

# Utilities
requests = "^2.31"
Faker = "^20.1"

[tool.poetry.group.dev.dependencies]
# Development and testing tools
pytest = "^7.4"
black = "^23.0"
ruff = "^0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Code formatter configuration (following KISS principle)
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | checkpoints
)/
'''

[tool.ruff]
# Fast Python linter (replaces flake8, isort, etc.)
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
]
exclude = [
    ".git",
    ".venv",
    "checkpoints",
    "__pycache__",
]
